Refactor: commit to new-style parsing
This replaces the previous font-tables crate with a new one, which will
be built from scratch on top of the new-style parse types.
