+++ font-types/src/raw.rs
@@ -32,6 +32,10 @@ pub trait ReadScalar: FixedSized {
 pub struct BigEndian<T: Scalar>(pub(crate) T::Raw);
 
 impl<T: Scalar> BigEndian<T> {
+    /// construct a new BigEndian<T> from raw bytes
+    pub fn new(raw: T::Raw) -> BigEndian<T> {
+        BigEndian(raw)
+    }
     /// Read a copy of this type from raw bytes.
     pub fn get(self) -> T {
         T::from_raw(self.0)
