+++ retired_crates/font-tables/Cargo.toml
@@ -7,14 +7,16 @@ license = "MIT/Apache-2.0"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [features]
 std = []
-
+compile = ["std"]
+diagnostics = ["diff", "ansi_term"]
+default = ["compile"]
 
 [dependencies]
 font-types = { path = "../font-types" }
+zerocopy = { version = "0.6.1" }
 bitflags = "1.3"
-#zerocopy = { version = "0.6.1" }
-#diff = { version = "0.1.12", optional = true }
-#ansi_term = { version = "0.12.1", optional = true }
+diff = { version = "0.1.12", optional = true }
+ansi_term = { version = "0.12.1", optional = true }
 
 [dev-dependencies]
 diff = "0.1.12"
@@ -22,6 +24,6 @@ xflags = "0.2.4"
 
 ansi_term = "0.12.1"
 
-#[[example]]
-#name = "gdef_check"
-#required-features = ["compile", "diagnostics"]
+[[example]]
+name = "gdef_check"
+required-features = ["compile", "diagnostics"]
