+++ retired_crates/font-tables/examples/inspect.rs
